makemove.py now checks for successful compilation, linking and packaging

+++ makemove.py
@@ -4,16 +4,19 @@ import shutil
 
 filename = "fx-CG50TETRIS.g3a"
 
-os.rename(filename, "__" + filename)
+if os.path.exists(filename):
+    os.rename(filename, "__" + filename)
 
 os.system("make")
 
-if os.path.exists("fx-CG50TETRIS.g3a"):
+if os.path.exists(filename):
     print("Build was successful")
-    os.remove("__" + filename)
+
+    if os.path.exists("__" + filename):
+        os.remove("__" + filename)
 
     if os.path.exists("E:"):
-        shutil.copyfile("fx-CG50TETRIS.g3a", "E:/fx-CG50TETRIS.g3a")
+        shutil.copyfile(filename, os.path.join("E:", filename))
         print("Move was successful")
     else:
         print("Drive not inserted, move not attempted")
@@ -21,4 +24,5 @@ if os.path.exists("fx-CG50TETRIS.g3a"):
 else:
     print("Build was unsuccessful, move not attempted")
 
-    os.rename("__" + filename, filename)
+    if os.path.exists("__" + filename):
+        os.rename("__" + filename, filename)
